backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
firecrawl-py==0.0.16
boto3==1.29.7
pdf2image==1.16.3
Pillow>=10.2.0
python-dotenv==1.0.0
pydantic>=2.10.0
requests==2.31.0
pypdf2==3.0.1

